pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "risk-monitor"
version = "0.1.0"
description = "Risk monitoring service for trading ecosystem"
authors = [
    {name = "Trading Ecosystem Team", email = "team@quantfidential.com"},
]
dependencies = [
    # FastAPI and HTTP server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Pydantic v2 with latest features
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # gRPC with async support and health checking
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-status>=1.68.0",
    "grpcio-health-checking>=1.68.0",
    "grpcio-reflection>=1.68.0",

    # Async utilities and concurrency
    "anyio>=4.6.0",
    "asyncio-mqtt>=0.16.0",

    # Logging and observability
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-grpc>=0.48b0",

    # Data persistence and caching
    "redis>=5.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",

    # HTTP client and utilities
    "httpx>=0.28.0",
    "tenacity>=9.0.0",

    # Data validation and serialization
    "orjson>=3.10.0",
    "msgpack>=1.1.0",
]
requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    # Testing framework
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-grpc>=0.8.0",

    # Code quality and formatting
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "black>=24.10.0",

    # gRPC testing utilities
    "grpcio-testing>=1.68.0",

    # Additional development tools
    "pre-commit>=4.0.0",
    "bandit>=1.8.0",
]

[tool.hatch.version]
path = "src/risk_monitor/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/risk_monitor"]

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/risk_monitor",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]